+++ PaymentHighway.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:Clutch.xcodeproj">
+      location = "self:/Users/binghuiliu/Documents/paymenthighway-ios-framework/PaymentHighway.xcodeproj">
    </FileRef>
 </Workspace>
